Fix subtle numpy upcast behavior causing bug. (#2991)
- Bug 1: Fix bug in quantization handling in runtime_torch_wrapper.py.
If you have a uint16-array and int32-scalar in numpy, then numpy doesn't
upcast, so we need to be explicit (note that in other array/scalar
combination this does happen, so it's very subtle).


Sorry, I claimed to have fixed this in
https://github.com/qcom-ai-hub/ai-hub-models-internal/pull/2984 in the
PR description, but I forgot to actually add the fix.

+++ src/qai_hub_models/utils/runtime_torch_wrapper.py
@@ -340,7 +340,10 @@ class RuntimeTorchWrapper(ABC, ExecutableModelProtocol, Generic[ModelIODetailsT]
                     and output_name in self.dequantize_model_output
                 ):
                     output = (
-                        (output - np.int32(output_details.qdq_params.zero_point))
+                        (
+                            output.astype(np.int32)
+                            - np.int32(output_details.qdq_params.zero_point)
+                        )
                         * np.float64(output_details.qdq_params.scale)
                     ).astype(np.float32)
